discovery: Move list of channels to a separate component
This will be reused in the Search page. Also for clarity: the new
component has only the state and logic needed to display the channels,
nothing else in the view.

https://phabricator.endlessm.com/T33030
