+++ kolibri_explore_plugin/assets/src/views/ChannelsPage.vue
@@ -37,23 +37,8 @@
       <template v-if="core.loading">
         <CardGridPlaceholder />
       </template>
-
       <template v-else>
-        <b-container class="channels pb-5">
-          <!-- Cards with thumbnail -->
-          <ChannelCardGroup
-            :rows="rows.withThumbnail"
-            :columns="columns"
-            @card-click="goToChannel"
-          />
-
-          <!-- Cards without thumbnail -->
-          <ChannelCardGroup
-            :rows="rows.withoutThumbnail"
-            :columns="columns"
-            @card-click="goToChannel"
-          />
-        </b-container>
+        <ChannelsList />
       </template>
 
     </div>
@@ -67,63 +52,24 @@
 
   import { mapState } from 'vuex';
   import commonCoreStrings from 'kolibri.coreVue.mixins.commonCoreStrings';
-  import _ from 'underscore';
-  import { responsiveMixin } from 'eos-components';
   import { PageNames, searchTerms } from '../constants';
-  import { RecommendedChannelIDs } from '../customApps';
 
+  import ChannelsList from '../components/ChannelsList';
   import DiscoveryNavBar from '../components/DiscoveryNavBar';
   import AboutModal from '../components/AboutModal';
 
   export default {
     name: 'ChannelsPage',
-    components: { AboutModal, DiscoveryNavBar },
-    mixins: [commonCoreStrings, responsiveMixin],
+    components: { AboutModal, ChannelsList, DiscoveryNavBar },
+    mixins: [commonCoreStrings],
     computed: {
-      ...mapState('topicsRoot', { channels: 'rootNodes', carouselNodes: 'carouselNodes' }),
+      ...mapState('topicsRoot', { carouselNodes: 'carouselNodes' }),
       ...mapState(['core']),
-      rows() {
-        let withThumbnail = [];
-        let withoutThumbnail = [];
-
-        this.channels.forEach(channel => {
-          if (RecommendedChannelIDs.includes(channel.id)) {
-            withThumbnail.push(channel);
-          } else {
-            withoutThumbnail.push(channel);
-          }
-        });
-
-        // Order the channels with thumbnail as in the ThumbApps array:
-        withThumbnail = _.sortBy(withThumbnail, n => RecommendedChannelIDs.indexOf(n.id));
-
-        // Split the channels in rows:
-        withThumbnail = _.chunk(withThumbnail, this.columns);
-        withoutThumbnail = _.chunk(withoutThumbnail, this.columns);
-        return { withThumbnail, withoutThumbnail };
-      },
-      columns() {
-        if (this.xs) {
-          return 1;
-        }
-
-        if (this.sm || this.md) {
-          return 2;
-        }
-
-        return 3;
-      },
       searchTerms() {
         return searchTerms;
       },
     },
     methods: {
-      goToChannel(channelId) {
-        this.$router.push({
-          name: PageNames.TOPICS_CHANNEL,
-          params: { channel_id: channelId },
-        });
-      },
       goToTerm(term) {
         const query = searchTerms.get(term) || term;
         this.$router.push({
@@ -152,10 +98,6 @@
 
   .channels-page {
     padding-top: $navbar-height;
-  }
-
-  .placeholder {
-    margin-top: $card-deck-margin * 2;
   }
 
 </style>
